Pluggable Cli Commands (#1938)
* make cli commands pluggable

* make deploy command behave correctly if not available

* add global debug flag

* * move plugin interface
* add tests for cli plugin discovery
* allow plugins to overwrite core cli commands

* ensure plugins take precedence over core commands
