+++ dlt/cli/__init__.py
@@ -1,0 +1,1 @@
+from dlt.cli.reference import SupportsCliCommand
